Cygwin fixes
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@300 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/d_monochromator.c
@@ -221,7 +221,7 @@ mxd_monochromator_get_enable_status( MX_RECORD *list_record,
 	*enable_status = *( int * ) pointer_to_value;
 
 	MX_DEBUG( 2,("%s: '%s' enable status = %d",
-		fname, list_record->name, *enable_status));
+		fname, list_record->name, (int) *enable_status));
 
 	return MX_SUCCESSFUL_RESULT;
 }
@@ -424,7 +424,7 @@ mxd_monochromator_restore_speeds( MX_MOTOR *motor )
 	for ( i = 0; i < monochromator->num_dependencies; i++ ) {
 
 		MX_DEBUG( 2,("%s: monochromator->speed_changed[%ld] = %d",
-			fname, i, monochromator->speed_changed[i]));
+			fname, i, (int) monochromator->speed_changed[i]));
 
 		if ( monochromator->speed_changed[i] ) {
