+++ libMx/i_compumotor.c
@@ -1076,7 +1076,7 @@ mxi_compumotor_multiaxis_move( MX_COMPUMOTOR_INTERFACE *compumotor_interface,
 
 		MX_DEBUG( 2,
 		("%s: i = %lu, will_be_moved = %d, destination[%lu] = %.5f",
-			fname, i, will_be_moved[i], i, destination[i]));
+			fname, i, (int) will_be_moved[i], i, destination[i]));
 
 		length = strlen( command );
